requirements.txt
anyio==3.6.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.9
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
tornado==6.2
tzdata==2022.1
tzlocal==4.2
